+++ mayan/apps/authentication_otp/tests/test_backends.py
@@ -30,7 +30,7 @@ class AuthenticationOTPBackendTestCase(
         ).tostr()
     )
     auto_login_user = False
-    create_test_case_superuser = True
+    create_test_case_super_user = True
 
     def setUp(self):
         super().setUp()
@@ -41,10 +41,10 @@ class AuthenticationOTPBackendTestCase(
         AuthenticationBackend.cls_initialize()
 
         user = authenticate(
-            username=self._test_case_superuser.email,
-            password=self._test_case_superuser.cleartext_password
+            username=self._test_case_super_user.email,
+            password=self._test_case_super_user.cleartext_password
         )
-        self.assertEqual(user, self._test_case_superuser)
+        self.assertEqual(user, self._test_case_super_user)
 
     @override_settings(AUTHENTICATION_BACKEND=PATH_AUTHENTICATION_BACKEND_EMAIL_OTP)
     def test_authentication_backend_email_with_otp(self):
@@ -53,26 +53,26 @@ class AuthenticationOTPBackendTestCase(
         self._enable_test_otp()
 
         user = authenticate(
-            username=self._test_case_superuser.email,
-            password=self._test_case_superuser.cleartext_password,
+            username=self._test_case_super_user.email,
+            password=self._test_case_super_user.cleartext_password,
         )
-        self.assertEqual(user, self._test_case_superuser)
+        self.assertEqual(user, self._test_case_super_user)
 
         user = authenticate(
             factor_name='otp_token', otp_token=self._test_token,
             user=user
         )
-        self.assertEqual(user, self._test_case_superuser)
+        self.assertEqual(user, self._test_case_super_user)
 
     @override_settings(AUTHENTICATION_BACKEND=PATH_AUTHENTICATION_BACKEND_USERNAME_OTP)
     def test_authentication_backend_username_no_otp(self):
         AuthenticationBackend.cls_initialize()
 
         user = authenticate(
-            username=self._test_case_superuser.username,
-            password=self._test_case_superuser.cleartext_password
+            username=self._test_case_super_user.username,
+            password=self._test_case_super_user.cleartext_password
         )
-        self.assertEqual(user, self._test_case_superuser)
+        self.assertEqual(user, self._test_case_super_user)
 
     @override_settings(AUTHENTICATION_BACKEND=PATH_AUTHENTICATION_BACKEND_USERNAME_OTP)
     def test_authentication_backend_username_with_otp(self):
@@ -81,13 +81,13 @@ class AuthenticationOTPBackendTestCase(
         self._enable_test_otp()
 
         user = authenticate(
-            username=self._test_case_superuser.username,
-            password=self._test_case_superuser.cleartext_password
+            username=self._test_case_super_user.username,
+            password=self._test_case_super_user.cleartext_password
         )
-        self.assertEqual(user, self._test_case_superuser)
+        self.assertEqual(user, self._test_case_super_user)
 
         user = authenticate(
             factor_name='otp_token', otp_token=self._test_token,
             user=user
         )
-        self.assertEqual(user, self._test_case_superuser)
+        self.assertEqual(user, self._test_case_super_user)