+++ mayan/apps/authentication_otp/tests/test_views.py
@@ -32,7 +32,7 @@ class LoginOTPTestCase(
         ).tostr()
     )
     auto_login_user = False
-    create_test_case_superuser = True
+    create_test_case_super_user = True
 
     def setUp(self):
         super().setUp()
@@ -54,13 +54,13 @@ class LoginOTPTestCase(
         self.assertEqual(events.count(), 2)
 
         self.assertEqual(events[0].action_object, None)
-        self.assertEqual(events[0].actor, self._test_case_superuser)
-        self.assertEqual(events[0].target, self._test_case_superuser)
+        self.assertEqual(events[0].actor, self._test_case_super_user)
+        self.assertEqual(events[0].target, self._test_case_super_user)
         self.assertEqual(events[0].verb, event_user_edited.id)
 
         self.assertEqual(events[1].action_object, None)
-        self.assertEqual(events[1].actor, self._test_case_superuser)
-        self.assertEqual(events[1].target, self._test_case_superuser)
+        self.assertEqual(events[1].actor, self._test_case_super_user)
+        self.assertEqual(events[1].target, self._test_case_super_user)
         self.assertEqual(events[1].verb, event_user_logged_in.id)
 
     @override_settings(AUTHENTICATION_BACKEND=PATH_AUTHENTICATION_BACKEND_EMAIL_OTP)
@@ -85,13 +85,13 @@ class LoginOTPTestCase(
         self.assertEqual(events.count(), 2)
 
         self.assertEqual(events[0].action_object, None)
-        self.assertEqual(events[0].actor, self._test_case_superuser)
-        self.assertEqual(events[0].target, self._test_case_superuser)
+        self.assertEqual(events[0].actor, self._test_case_super_user)
+        self.assertEqual(events[0].target, self._test_case_super_user)
         self.assertEqual(events[0].verb, event_user_edited.id)
 
         self.assertEqual(events[1].action_object, None)
-        self.assertEqual(events[1].actor, self._test_case_superuser)
-        self.assertEqual(events[1].target, self._test_case_superuser)
+        self.assertEqual(events[1].actor, self._test_case_super_user)
+        self.assertEqual(events[1].target, self._test_case_super_user)
         self.assertEqual(events[1].verb, event_user_logged_in.id)
 
     @override_settings(AUTHENTICATION_BACKEND=PATH_AUTHENTICATION_BACKEND_USERNAME_OTP)
@@ -110,13 +110,13 @@ class LoginOTPTestCase(
         self.assertEqual(events.count(), 2)
 
         self.assertEqual(events[0].action_object, None)
-        self.assertEqual(events[0].actor, self._test_case_superuser)
-        self.assertEqual(events[0].target, self._test_case_superuser)
+        self.assertEqual(events[0].actor, self._test_case_super_user)
+        self.assertEqual(events[0].target, self._test_case_super_user)
         self.assertEqual(events[0].verb, event_user_edited.id)
 
         self.assertEqual(events[1].action_object, None)
-        self.assertEqual(events[1].actor, self._test_case_superuser)
-        self.assertEqual(events[1].target, self._test_case_superuser)
+        self.assertEqual(events[1].actor, self._test_case_super_user)
+        self.assertEqual(events[1].target, self._test_case_super_user)
         self.assertEqual(events[1].verb, event_user_logged_in.id)
 
     @override_settings(AUTHENTICATION_BACKEND=PATH_AUTHENTICATION_BACKEND_USERNAME_OTP)
@@ -141,13 +141,13 @@ class LoginOTPTestCase(
         self.assertEqual(events.count(), 2)
 
         self.assertEqual(events[0].action_object, None)
-        self.assertEqual(events[0].actor, self._test_case_superuser)
-        self.assertEqual(events[0].target, self._test_case_superuser)
+        self.assertEqual(events[0].actor, self._test_case_super_user)
+        self.assertEqual(events[0].target, self._test_case_super_user)
         self.assertEqual(events[0].verb, event_user_edited.id)
 
         self.assertEqual(events[1].action_object, None)
-        self.assertEqual(events[1].actor, self._test_case_superuser)
-        self.assertEqual(events[1].target, self._test_case_superuser)
+        self.assertEqual(events[1].actor, self._test_case_super_user)
+        self.assertEqual(events[1].target, self._test_case_super_user)
         self.assertEqual(events[1].verb, event_user_logged_in.id)
 
     @override_settings(AUTHENTICATION_BACKEND=PATH_AUTHENTICATION_BACKEND_USERNAME_OTP)
@@ -187,13 +187,13 @@ class LoginOTPTestCase(
         self.assertEqual(events.count(), 2)
 
         self.assertEqual(events[0].action_object, None)
-        self.assertEqual(events[0].actor, self._test_case_superuser)
-        self.assertEqual(events[0].target, self._test_case_superuser)
+        self.assertEqual(events[0].actor, self._test_case_super_user)
+        self.assertEqual(events[0].target, self._test_case_super_user)
         self.assertEqual(events[0].verb, event_user_edited.id)
 
         self.assertEqual(events[1].action_object, None)
-        self.assertEqual(events[1].actor, self._test_case_superuser)
-        self.assertEqual(events[1].target, self._test_case_superuser)
+        self.assertEqual(events[1].actor, self._test_case_super_user)
+        self.assertEqual(events[1].target, self._test_case_super_user)
         self.assertEqual(events[1].verb, event_user_logged_in.id)