+++ mayan/apps/locales/tests/mixins.py
@@ -14,14 +14,14 @@ class UserLocaleProfileViewMixin:
             user=self._test_case_user, follow=follow
         )
 
-    def _request_test_superuser_locale_profile_detail_view(self):
+    def _request_test_super_user_locale_profile_detail_view(self):
         return self._request_test_user_locale_profile_detail_view(
-            user=self._test_superuser
+            user=self._test_super_user
         )
 
-    def _request_test_superuser_locale_profile_edit_view(self):
+    def _request_test_super_user_locale_profile_edit_view(self):
         return self._request_test_user_locale_profile_edit_view(
-            user=self._test_superuser
+            user=self._test_super_user
         )
 
     def _request_test_user_locale_profile_detail_view(self, user=None):