+++ mayan/apps/locales/tests/test_views.py
@@ -53,69 +53,69 @@ class SuperUserLocaleViewTestCase(
 ):
     def setUp(self):
         super().setUp()
-        self._create_test_superuser()
+        self._create_test_super_user()
 
-    def test_superuser_locale_profile_detail_view_no_permission(self):
+    def test_super_user_locale_profile_detail_view_no_permission(self):
         self._clear_events()
 
-        response = self._request_test_superuser_locale_profile_detail_view()
+        response = self._request_test_super_user_locale_profile_detail_view()
         self.assertEqual(response.status_code, 404)
 
         events = self._get_test_events()
         self.assertEqual(events.count(), 0)
 
-    def test_superuser_locale_profile_detail_view_with_access(self):
+    def test_super_user_locale_profile_detail_view_with_access(self):
         self.grant_access(
-            obj=self._test_superuser, permission=permission_user_view
+            obj=self._test_super_user, permission=permission_user_view
         )
 
         self._clear_events()
 
-        response = self._request_test_superuser_locale_profile_detail_view()
+        response = self._request_test_super_user_locale_profile_detail_view()
         self.assertEqual(response.status_code, 404)
 
         events = self._get_test_events()
         self.assertEqual(events.count(), 0)
 
-    def test_superuser_locale_profile_edit_view_no_permission(self):
-        language = self._test_superuser.locale_profile.language
-        timezone = self._test_superuser.locale_profile.timezone
+    def test_super_user_locale_profile_edit_view_no_permission(self):
+        language = self._test_super_user.locale_profile.language
+        timezone = self._test_super_user.locale_profile.timezone
 
         self._clear_events()
 
-        response = self._request_test_superuser_locale_profile_edit_view()
+        response = self._request_test_super_user_locale_profile_edit_view()
         self.assertEqual(response.status_code, 404)
 
-        self._test_superuser.refresh_from_db()
+        self._test_super_user.refresh_from_db()
         self.assertEqual(
-            self._test_superuser.locale_profile.language, language
+            self._test_super_user.locale_profile.language, language
         )
         self.assertEqual(
-            self._test_superuser.locale_profile.timezone, timezone
+            self._test_super_user.locale_profile.timezone, timezone
         )
 
         events = self._get_test_events()
         self.assertEqual(events.count(), 0)
 
-    def test_superuser_locale_profile_edit_view_with_access(self):
-        language = self._test_superuser.locale_profile.language
-        timezone = self._test_superuser.locale_profile.timezone
+    def test_super_user_locale_profile_edit_view_with_access(self):
+        language = self._test_super_user.locale_profile.language
+        timezone = self._test_super_user.locale_profile.timezone
 
         self.grant_access(
-            obj=self._test_superuser, permission=permission_user_edit
+            obj=self._test_super_user, permission=permission_user_edit
         )
 
         self._clear_events()
 
-        response = self._request_test_superuser_locale_profile_edit_view()
+        response = self._request_test_super_user_locale_profile_edit_view()
         self.assertEqual(response.status_code, 404)
 
-        self._test_superuser.refresh_from_db()
+        self._test_super_user.refresh_from_db()
         self.assertEqual(
-            self._test_superuser.locale_profile.language, language
+            self._test_super_user.locale_profile.language, language
         )
         self.assertEqual(
-            self._test_superuser.locale_profile.timezone, timezone
+            self._test_super_user.locale_profile.timezone, timezone
         )
 
         events = self._get_test_events()