+++ mayan/apps/quotas/tests/test_quotas.py
@@ -372,8 +372,8 @@ class DocumentCountQuotaTestCase(GroupTestMixin, GenericDocumentTestCase):
             events[10].verb, event_document_version_page_created.id
         )
 
-    def test_superuser_restriction(self):
-        self._create_test_superuser()
+    def test_super_user_restriction(self):
+        self._create_test_super_user()
 
         self._test_quota = DocumentCountQuota.create(
             documents_limit=1,
@@ -384,7 +384,7 @@ class DocumentCountQuotaTestCase(GroupTestMixin, GenericDocumentTestCase):
             user_ids=()
         )
 
-        self._upload_test_document(user=self._test_superuser)
+        self._upload_test_document(user=self._test_super_user)
 
         events = self._get_test_events()
         self.assertEqual(events.count(), 12)
@@ -421,8 +421,8 @@ class DocumentCountQuotaTestCase(GroupTestMixin, GenericDocumentTestCase):
         )
 
         self.assertEqual(events[5].action_object, None)
-        self.assertEqual(events[5].actor, self._test_superuser)
-        self.assertEqual(events[5].target, self._test_superuser)
+        self.assertEqual(events[5].actor, self._test_super_user)
+        self.assertEqual(events[5].target, self._test_super_user)
         self.assertEqual(events[5].verb, event_user_created.id)
 
         self.assertEqual(events[6].action_object, None)
@@ -431,29 +431,29 @@ class DocumentCountQuotaTestCase(GroupTestMixin, GenericDocumentTestCase):
         self.assertEqual(events[6].verb, event_quota_created.id)
 
         self.assertEqual(events[7].action_object, self._test_document_type)
-        self.assertEqual(events[7].actor, self._test_superuser)
+        self.assertEqual(events[7].actor, self._test_super_user)
         self.assertEqual(events[7].target, self._test_document)
         self.assertEqual(events[7].verb, event_document_created.id)
 
         self.assertEqual(events[8].action_object, self._test_document)
-        self.assertEqual(events[8].actor, self._test_superuser)
+        self.assertEqual(events[8].actor, self._test_super_user)
         self.assertEqual(events[8].target, self._test_document_file)
         self.assertEqual(events[8].verb, event_document_file_created.id)
 
         self.assertEqual(events[9].action_object, self._test_document)
-        self.assertEqual(events[9].actor, self._test_superuser)
+        self.assertEqual(events[9].actor, self._test_super_user)
         self.assertEqual(events[9].target, self._test_document_file)
         self.assertEqual(events[9].verb, event_document_file_edited.id)
 
         self.assertEqual(events[10].action_object, self._test_document)
-        self.assertEqual(events[10].actor, self._test_superuser)
+        self.assertEqual(events[10].actor, self._test_super_user)
         self.assertEqual(events[10].target, self._test_document_version)
         self.assertEqual(events[10].verb, event_document_version_created.id)
 
         self.assertEqual(
             events[11].action_object, self._test_document_version
         )
-        self.assertEqual(events[11].actor, self._test_superuser)
+        self.assertEqual(events[11].actor, self._test_super_user)
         self.assertEqual(events[11].target, self._test_document_version_page)
         self.assertEqual(
             events[11].verb, event_document_version_page_created.id
@@ -609,8 +609,8 @@ class DocumentSizeQuotaTestCase(GroupTestMixin, GenericDocumentTestCase):
             events[4].verb, event_document_version_page_created.id
         )
 
-    def test_superuser_restriction(self):
-        self._create_test_superuser()
+    def test_super_user_restriction(self):
+        self._create_test_super_user()
 
         self._test_quota = DocumentSizeQuota.create(
             document_size_limit=0.01,
@@ -623,35 +623,35 @@ class DocumentSizeQuotaTestCase(GroupTestMixin, GenericDocumentTestCase):
 
         self._clear_events()
 
-        self._upload_test_document(user=self._test_superuser)
+        self._upload_test_document(user=self._test_super_user)
 
         events = self._get_test_events()
         self.assertEqual(events.count(), 5)
 
         self.assertEqual(events[0].action_object, self._test_document_type)
-        self.assertEqual(events[0].actor, self._test_superuser)
+        self.assertEqual(events[0].actor, self._test_super_user)
         self.assertEqual(events[0].target, self._test_document)
         self.assertEqual(events[0].verb, event_document_created.id)
 
         self.assertEqual(events[1].action_object, self._test_document)
-        self.assertEqual(events[1].actor, self._test_superuser)
+        self.assertEqual(events[1].actor, self._test_super_user)
         self.assertEqual(events[1].target, self._test_document_file)
         self.assertEqual(events[1].verb, event_document_file_created.id)
 
         self.assertEqual(events[2].action_object, self._test_document)
-        self.assertEqual(events[2].actor, self._test_superuser)
+        self.assertEqual(events[2].actor, self._test_super_user)
         self.assertEqual(events[2].target, self._test_document_file)
         self.assertEqual(events[2].verb, event_document_file_edited.id)
 
         self.assertEqual(events[3].action_object, self._test_document)
-        self.assertEqual(events[3].actor, self._test_superuser)
+        self.assertEqual(events[3].actor, self._test_super_user)
         self.assertEqual(events[3].target, self._test_document_version)
         self.assertEqual(events[3].verb, event_document_version_created.id)
 
         self.assertEqual(
             events[4].action_object, self._test_document_version
         )
-        self.assertEqual(events[4].actor, self._test_superuser)
+        self.assertEqual(events[4].actor, self._test_super_user)
         self.assertEqual(events[4].target, self._test_document_version_page)
         self.assertEqual(
             events[4].verb, event_document_version_page_created.id